Move xml builders into lib
